+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## et-orbi 1.3.0  not yet released
 
+- Introduce EtOrbi.zone_name_abbreviation(zone_name, time, dst=false)
+
 
 ## et-orbi 1.2.11  released 2024-03-23